Resolve review feedback
