+++ apps/web/.eslintrc.json
@@ -6,7 +6,7 @@
     "no-restricted-imports": [
       "error",
       {
-        "patterns": ["**/core/*"]
+        "patterns": ["**/core/*", "**/reports/*"]
       }
     ]
   }
