+++ apps/web/src/app/reports/index.ts
@@ -1,1 +1,3 @@
 export * from "./reports.module";
+export * from "./models/report-entry";
+export * from "./models/report-variant";
