+++ apps/web/src/app/reports/models/report-entry.ts
@@ -1,0 +1,9 @@
+import { ReportVariant } from "./report-variant";
+
+export type ReportEntry = {
+  title: string;
+  description: string;
+  route: string;
+  icon: string;
+  variant: ReportVariant;
+};
