+++ apps/web/src/app/reports/models/report-variant.ts
@@ -1,0 +1,5 @@
+export enum ReportVariant {
+  Enabled = "Enabled",
+  RequiresPremium = "RequiresPremium",
+  RequiresUpgrade = "RequiresUpgrade",
+}
