+++ apps/web/src/app/reports/pages/reports-home.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from "@angular/core";
 
 import { StateService } from "@bitwarden/common/abstractions/state.service";
 
-import { ReportVariant } from "../report-card/report-card.component";
-import { ReportEntry } from "../report-list/report-list.component";
+import { ReportEntry } from "../models/report-entry";
+import { ReportVariant } from "../models/report-variant";
 import { reports, ReportType } from "../reports";
 
 @Component({
